Made undo file more intuitive and added a confirmation modal, merged git config modal with settings modal
