+++ git/GitClient.ts
@@ -353,10 +353,21 @@ export class GitClient {
   }
 
   undoFileChanges(file: string, revision: string, staged: boolean): Promise<any> {
-    return this.simpleOperation(
-      this.getGitPath(),
-      ['checkout', '-q', (revision || 'HEAD'), '--', file],
-      'Undo File Changes');
+    if (staged) {
+      return this.simpleOperation(
+        this.getGitPath(),
+        ['checkout', '-q', (revision || 'HEAD'), '--', file],
+        'Undo File Changes');
+    } else {
+      return this.simpleOperation(
+        this.getGitPath(),
+        ['stash', 'save', '--keep-index', '--', file],
+        'Stash Local File Changes').then(() => this.simpleOperation(
+        this.getGitPath(),
+        ['stash', 'drop'],
+        'Drop Stashed Local File Changes',
+      ));
+    }
   }
 
   hardReset(): Promise<any> {
