--- src/app/components/git-config/git-config.component.html
@@ -1,61 +1,0 @@
-<button class="btn btn-secondary center-text-vertical mx-2" title="Git Config"
-        (click)="show()">
-  <i class="fa fa-wrench"></i>
-  <span class="ml-1">Config</span>
-</button>
-<app-modal modalId="gitConfig"
-           modalTitle="Git Config"
-           modalTitleIcon="fa-wrench"
-           modalClass="dialog-wide"
-           negativeButtonText="Close"
-           [hasAffirmative]="false"
-           (onCancel)="close()">
-  <div></div>
-  <input class="form-control" [(ngModel)]="filter" placeholder="Filter...">
-  <table class="table table-striped">
-    <thead>
-    <tr>
-      <th>Key</th>
-      <th>Value</th>
-      <th>Location</th>
-      <th width="1%"></th>
-    </tr>
-    </thead>
-    <tbody>
-    <tr *ngFor="let item of items | filterObject : 'key' : filter">
-      <td class="editable-item" (click)="startEdit(item,true)">
-        <span *ngIf="!isEditing(item) || !clickedKey">{{item.key}}</span>
-        <input class="form-control"
-               *ngIf="isEditing(item) && clickedKey"
-               (keyup.enter)="saveConfigItem(item,item)"
-               (blur)="saveConfigItem(item,item)"
-               (keydown.tab)="clickedKey=false"
-               (keydown.escape)="cancelEdit()"
-               [(ngModel)]="editedItem.key"
-               [autofocus]>
-      </td>
-      <td class="editable-item" (click)="startEdit(item,false)">
-        <span *ngIf="!isEditing(item)|| clickedKey">{{item.value}}</span>
-        <input class="form-control"
-               *ngIf="isEditing(item) && !clickedKey"
-               (keyup.enter)="saveConfigItem(item)"
-               (blur)="saveConfigItem(item)"
-               (keydown.shift.tab)="clickedKey=true"
-               (keydown.tab)="nextTabRow(item)"
-               (keydown.escape)="cancelEdit()"
-               [(ngModel)]="editedItem.value"
-               [autofocus]>
-      </td>
-      <td>{{getConfigFileDisplay(item.sourceFile)}}</td>
-      <td>
-        <button class="table-btn btn btn-outline-danger" (click)="deleteConfigItem(item)">
-          <i class="fa fa-trash"></i>
-        </button>
-      </td>
-    </tr>
-    <tr class="editable-item" (click)="newItem()">
-      <td class="text-muted" colspan="4">New Item...</td>
-    </tr>
-    </tbody>
-  </table>
-</app-modal>
