--- src/app/components/git-config/git-config.component.scss
@@ -1,10 +1,0 @@
-.editable-item {
-  cursor: pointer;
-  &:hover {
-    background-color: rgba(128, 128, 128, .3);
-  }
-}
-
-.table-btn {
-  margin: -.5em;
-}
